back end Tests done

+++ Bank200/Service/IAccountService.cs
@@ -4,11 +4,11 @@
     {
         public void openSavingsAccount(long accountId, long amountToDeposit);
         public void openCurrentAccount(long accountId);
-        public void withdraw(long accountId, int amountToWithdraw);
+        public void withdraw(long accountId, long amountToWithdraw);
 
         // throw AccountNotFoundException, WithdrawalAmountTooLargeException; 
 
-        public void deposit(long accountId, int amountToDeposit);
+        public void deposit(long accountId, long amountToDeposit);
 
         // throw AccountNotFoundException;
     }
